:truck: move methods to sub folder
