+++ test/apply-key.js
@@ -1,5 +1,5 @@
 import test from 'ava';
-import {applyKey} from '../index.js';
+import {applyKey} from '../src/cypher.js';
 
 test('abc on UPPERCASE', t => {
 	t.is(applyKey('THE QUICK BROWN FOX JUMPS OVER THE LAZY DOG.', [0, 1, 2]), 'TIG QVKCL DRPYN GQX KWMQU OWGR UJE MCZZ FOH.');
